Subtracted time required for communication between switches and controller, when computing delay for VoIP

+++ impl/src/main/java/sqmf/impl/VoIP.java
@@ -33,7 +33,7 @@ public class VoIP {
      * @param packetLoss    The computed packet loss (%).
      * @return              The computed QoE value.
      */
-    public static double estimateQoE(Long delay, double packetLoss) {
+    public static double estimateQoE(float delay, double packetLoss) {
         int h;
         if (delay - 177.3 > 0){
             h = 1;
